requirements.txt
Django~=2.2.4
django-ckeditor~=5.9.0